+++ morn-core/src/main/java/site/morn/rest/RestPage.java
@@ -1,8 +1,7 @@
-package site.morn.boot.rest;
+package site.morn.rest;
 
 import io.swagger.annotations.ApiModel;
 import site.morn.core.CriteriaMap;
-import site.morn.rest.RestModel;
 import site.morn.util.GenericUtils;
 
 /**
